Готово

+++ database/database.py
@@ -1,7 +1,46 @@
+import os
+
 import aiosqlite
 import json
 import pprint
 
+
+# Создание БД для квизов
+async def crete_quizzes():
+    async with aiosqlite.connect('userdata.db') as db:
+        await db.execute(
+            'CREATE TABLE IF NOT EXISTS quizzes (id_quizzes INTEGER PRIMARY KEY,'
+            'name_quiz TEXT, one_question TEXT, one_answer TEXT, two_question TEXT, two_answer TEXT,'
+            'three_question TEXT, three_answer TEXT, prize TEXT)'
+        )
+        await db.commit()
+
+# Добавление в БД новых квизов
+async def append_new_quizzes(name_quizzes, one_question, one_answer, two_question, two_answer,
+                             three_question, three_answer, prize):
+    async with aiosqlite.connect('userdata.db') as db:
+        await db.execute(
+            'INSERT OR IGNORE INTO quizzes (name_quiz, one_question, one_answer, two_question,'
+            'two_answer, three_question, three_answer, prize)'
+            'VALUES (?, ?, ?, ?, ?, ?, ?, ?)', (name_quizzes, one_question, one_answer,
+                                                two_question, two_answer, three_question, three_answer, prize)
+        )
+        await db.commit()
+
+async def create_task_table():
+    async with aiosqlite.connect('userdata.db') as db:
+        await db.execute('CREATE TABLE IF NOT EXISTS task (id INTEGER PRIMARY KEY, url TEXT, bonus TEXT)')
+        await db.commit()
+# Запись БД в json-файл
+async def info_quizzes_for_quizzes_json():
+    async with aiosqlite.connect('userdata.db') as con:
+        cursor = await con.execute(f'SELECT * FROM quizzes')
+        data = await cursor.fetchall()
+    exchange_data_quizzes = []
+    for row in data:
+        exchange_data_quizzes.append({str(row[1]): {str(row[2]): str(row[3]), str(row[4]): str(row[5]), str(row[6]): str(row[7]), "награда": str(row[8])}})
+    with open('exchange_look_quizzes.json', 'w', encoding='utf-8') as json_file:
+        json.dump(exchange_data_quizzes, json_file, ensure_ascii=False, indent=4)
 
 # Создание БД для отложенных сообщений
 async def create_old_message():
@@ -63,7 +102,7 @@ async def info_from_old_message_in_json():
     async with aiosqlite.connect('userdata.db') as con:
         cursor = await con.execute(f'SELECT * FROM old_message')
         data = await cursor.fetchall()
-        await con.close()
+        await con.commit()
     exchange_data_old = []
     for row in data:
         exchange_data_old.append({
@@ -119,6 +158,14 @@ async def create_admin():
         )
         await db.commit()
 
+async def create_merch():
+    async with aiosqlite.connect('userdata.db') as db:
+        await db.execute('CREATE TABLE IF NOT EXISTS merch (id INTEGER PRIMARY KEY AUTOINCREMENT,'
+                         'name VARCHAR(255),'
+                         'price DECIMAL(10, 2),'
+                         'image_path VARCHAR(255))')
+        await db.commit()
+
 
 # Добавление данных в БД admin
 async def add_admin_to_admin_table(admin_id: str, username: str, Lucky: float, CashOnline: float,
@@ -155,7 +202,7 @@ async def create_users():
         await db.execute(
             "CREATE TABLE IF NOT EXISTS users (user_id INTEGER PRIMARY KEY, "
             "username TEXT, nickname TEXT, Lucky REAL, CashOnline REAL, OtherCoins Real,"
-            "Chips Real, referral TEXT, current_time TIMESTAMP)"
+            "Chips Real, referral TEXT DEFAULT '[]', current_time TIMESTAMP)"
         )
         await db.commit()
 
@@ -218,35 +265,38 @@ async def add_user_to_users_table(user_id: str, username: str, nickname, Lucky: 
         await db.execute(
             "INSERT OR IGNORE INTO users (user_id, username, nickname, Lucky, CashOnline, OtherCoins, Chips, current_time)"
             "VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
-            [user_id, username, nickname, Lucky, CashOnline, OtherCoins, Chips, current_time]
+            [user_id, username, nickname, Lucky, CashOnline, OtherCoins, Chips,  current_time]
         )
         await db.commit()
 
 
 # Сохраняем базу юзеров в json-файл
-async def info_user_for_user(current_time):
+async def info_user_for_user(user_id, current_time):
     async with aiosqlite.connect('userdata.db') as con:
-        cursor = await con.execute(F'SELECT * FROM users')
+        cursor = await con.execute('SELECT * FROM users WHERE user_id = ?', (user_id,))
         data = await cursor.fetchall()
     exchange_data = []
     for row in data:
         exchange_data.append({
-            'id': row[0],
-            'username': row[1],
-            'nickname': row[2],
-            'Lucky': row[3],
-            'CashOnline': row[4],
-            'OtherCoins': row[5],
-            'Chips': row[6],
-            'referral': row[7],
-            'current_time': current_time
+            "id": row[0],
+            "username": row[1],
+            "nickname": row[2],
+            "Lucky": row[3],
+            "CashOnline": row[4],
+            "OtherCoins": row[5],
+            "Chips": row[6],
+            "referral": row[7],
+            "current_time": current_time
         })
-    with open('exchange_data.json', 'r', encoding='utf-8') as file:
-        data = json.load(file)
-    data.append(exchange_data[0])
-    with open('exchange_data.json', 'w', encoding='utf-8') as json_file:
-        json.dump(data, json_file, ensure_ascii=False, indent=4)
-
+    if os.path.exists('exchange_data.json') == True:
+        with open('exchange_data.json', 'r', encoding='utf-8') as file:
+            data = json.load(file)
+        data.append(exchange_data[0])
+        with open('exchange_data.json', 'w', encoding='utf-8') as json_file:
+            json.dump(data, json_file, ensure_ascii=False, indent=4)
+    else:
+        with open('exchange_data.json', 'w', encoding='utf-8') as json_file:
+            json.dump(exchange_data, json_file, ensure_ascii=False, indent=4)
 
 async def give_bonus(count, user_id):
     async with aiosqlite.connect('userdata.db') as conn:
